+++ src/Orc.Controls/Controls/TimePicker/TimePickerInputController.cs
@@ -3,10 +3,10 @@
     using System;
     using System.Windows;
     using static Orc.Controls.ClockMath;
-    using static Orc.Controls.ClockFace;
-    public class ClockFaceInputController
+    using static Orc.Controls.TimePicker;
+    public class TimePickerInputController
     {
-        private readonly ClockFace _clockFace;
+        private readonly TimePicker _timePicker;
 
         // TimePicker.ActualHeight * MinDistanceRatio is the max
         // distance away from the tip of the indicator you can 
@@ -15,22 +15,22 @@
 
         private Indicator _indicator;
         private bool _isDragging;
-        public ClockFaceInputController(ClockFace clockFace)
+        public TimePickerInputController(TimePicker timePicker)
         {
-            _clockFace = clockFace;
+            _timePicker = timePicker;
 
-            _clockFace.PreviewMouseLeftButtonDown += _clockFace_PreviewMouseLeftButtonDown;
-            _clockFace.PreviewMouseMove += _clockFace_PreviewMouseMove;
-            _clockFace.MouseLeave += _clockFace_MouseLeave; ;
-            _clockFace.PreviewMouseLeftButtonUp += _clockFace_PreviewMouseLeftButtonUp; ;
+            _timePicker.PreviewMouseLeftButtonDown += _timePicker_PreviewMouseLeftButtonDown;
+            _timePicker.PreviewMouseMove += _timePicker_PreviewMouseMove;
+            _timePicker.MouseLeave += _timePicker_MouseLeave; ;
+            _timePicker.PreviewMouseLeftButtonUp += _timePicker_PreviewMouseLeftButtonUp; ;
         }
 
 
         private void StartDragging(Point mouse)
         {
-            var width = _clockFace.ActualWidth;
-            var height = _clockFace.ActualHeight;
-            var radius = (Math.Min(width, height) - _clockFace.BorderThickness.Left) / 2.0;
+            var width = _timePicker.ActualWidth;
+            var height = _timePicker.ActualHeight;
+            var radius = (Math.Min(width, height) - _timePicker.BorderThickness.Left) / 2.0;
             var center = new Point(width / 2.0, height / 2.0);
 
             // TODO: highlight indicator that you're dragging
@@ -45,36 +45,36 @@
             _isDragging = false;
         }
 
-        private void _clockFace_PreviewMouseMove(object sender, System.Windows.Input.MouseEventArgs e)
+        private void _timePicker_PreviewMouseMove(object sender, System.Windows.Input.MouseEventArgs e)
         {
             if (_isDragging)
             {
-                var width = _clockFace.ActualWidth;
-                var height = _clockFace.ActualHeight;
-                var radius = (Math.Min(width, height) - _clockFace.BorderThickness.Left) / 2.0;
+                var width = _timePicker.ActualWidth;
+                var height = _timePicker.ActualHeight;
+                var radius = (Math.Min(width, height) - _timePicker.BorderThickness.Left) / 2.0;
                 var center = new Point(width / 2.0, height / 2.0);
-                var mouse = e.GetPosition(_clockFace);
+                var mouse = e.GetPosition(_timePicker);
 
-                var time = _clockFace.TimeValue;
+                var time = _timePicker.TimeValue;
 
                 if (_indicator == Indicator.HourIndicator)
                 {
                     var hour = AngleToHour(center, mouse);
-                    _clockFace.SetCurrentValue(TimeValueProperty, new TimeSpan(hour, time.Minutes, time.Seconds));
+                    _timePicker.SetCurrentValue(TimeValueProperty, new TimeSpan(hour, time.Minutes, time.Seconds));
                 }
 
                 if (_indicator == Indicator.MinuteIndicator)
                 {
                     var minutes = AngleToMinutes(center, mouse);
-                    _clockFace.SetCurrentValue(TimeValueProperty, new TimeSpan(time.Hours, minutes, time.Seconds));
+                    _timePicker.SetCurrentValue(TimeValueProperty, new TimeSpan(time.Hours, minutes, time.Seconds));
                 }
             }
         }
 
         private void FindIndicator(double width, double height, double radius, Point center, Point mouse)
         {
-            var minuteTip = LineOnCircle((Math.PI * 2 * _clockFace.TimeValue.Minutes / 60) - Math.PI / 2.0, center, 0, radius * MinuteIndicatorRatio)[1];
-            var hourTip = LineOnCircle((Math.PI * 2 * _clockFace.TimeValue.Hours / 12) - Math.PI / 2.0, center, 0, radius * HourIndicatorRatio)[1];
+            var minuteTip = LineOnCircle((Math.PI * 2 * _timePicker.TimeValue.Minutes / 60) - Math.PI / 2.0, center, 0, radius * MinuteIndicatorRatio)[1];
+            var hourTip = LineOnCircle((Math.PI * 2 * _timePicker.TimeValue.Hours / 12) - Math.PI / 2.0, center, 0, radius * HourIndicatorRatio)[1];
 
             var maxDistance = width * MinDistanceRatio;
 
@@ -91,17 +91,17 @@
             }
         }
 
-        private void _clockFace_PreviewMouseLeftButtonDown(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        private void _timePicker_PreviewMouseLeftButtonDown(object sender, System.Windows.Input.MouseButtonEventArgs e)
         {
-            StartDragging(e.GetPosition(_clockFace));
+            StartDragging(e.GetPosition(_timePicker));
         }
 
-        private void _clockFace_PreviewMouseLeftButtonUp(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        private void _timePicker_PreviewMouseLeftButtonUp(object sender, System.Windows.Input.MouseButtonEventArgs e)
         {
             StopDragging();
         }
 
-        private void _clockFace_MouseLeave(object sender, System.Windows.Input.MouseEventArgs e)
+        private void _timePicker_MouseLeave(object sender, System.Windows.Input.MouseEventArgs e)
         {
             StopDragging();
         }
